move getInitialState method from utility.js to App.js
